Some more refactoring around evaluations.py
Note that there is no new net behavior
